Update Bullet.cs
Adds logic to support pooling the game object by cycling through SetActive On Enable.

In other words, when the bullet is enabled, the OnEnable function Invokes the `Hide` function for 1 second.
Hide simply SetsActive to false which is then switched to true when you press space bar. That logic is in the PoolManager script.